Update bulk-upsert schema to accept floats for stock fields and align tip_produs_id_sub usage
- Change ro_stock and hu_stock in schema from integer to number (float)
- Ensure ProductUpsertItem accepts numeric values for these fields
- Prepare for syncing tip_produs_id_sub without type mismatch errors

X-Lovable-Edit-ID: edt-1589f5ac-a34a-4063-acd0-dbe77f3c2fe5

+++ supabase/functions/_shared/schema.ts
@@ -55,9 +55,9 @@ export const ProductUpsertItem = z.object({
   senior_erp_link: z.string().nullable().optional(),
   site_ro_product_id: z.number().int().nullable().optional(),
   site_hu_product_id: z.number().int().nullable().optional(),
-  ro_stock: z.number().int().nullable().optional(),
+  ro_stock: z.number().nullable().optional(),
   ro_stoc_detailed: z.string().nullable().optional(),
-  hu_stock: z.number().int().nullable().optional(),
+  hu_stock: z.number().nullable().optional(),
   hu_stock_detailed: z.string().nullable().optional(),
   tip_produs_id_sub: z.number().int().nullable().optional(),
   tip_produs_id_main: z.number().int().nullable().optional()
